+++ seiscat/scripts/seiscat.py
@@ -36,6 +36,15 @@ def run():
     elif args.action == 'editdb':
         from ..editdb import editdb
         editdb(config)
+    elif args.action == 'download':
+        download_event = args.event or args.both
+        download_data = args.data or args.both
+        if download_event:
+            from ..download_event_details import download_event_details
+            download_event_details(config)
+        if download_data:
+            from ..download_event_waveforms import download_event_waveforms
+            download_event_waveforms(config)
     elif args.action == 'print':
         from ..print import print_catalog
         print_catalog(config)